+++ docs/src/routes/+layout.svelte
@@ -1,13 +1,14 @@
 <script lang="ts">
 	import { PUBLIC_BASE_PATH } from '$env/static/public';
 	import './layout.scss';
-	import { SvelAside } from '@svelement-ui/all';
-	import SvelContainer from '@svelement-ui/container';
-	import SvelHeader from '@svelement-ui/header';
-	import SvelMain from '@svelement-ui/main';
-	import SvelFooter from '@svelement-ui/footer';
-	import SvelRow from '@svelement-ui/row';
-	import SvelCol from '@svelement-ui/col';
+	import {
+		SvelCol,
+		SvelContainer,
+		SvelFooter,
+		SvelHeader,
+		SvelMain,
+		SvelRow,
+	} from '@svelement-ui/all';
 </script>
 
 <SvelContainer direction="vertical">