+++ packages/all/src/index.ts
@@ -1,12 +1,12 @@
 import '@svelement-ui/theme-chalk/src/index.scss';
 
-import SvelAside from '@svelement-ui/aside/src/svel-aside.svelte';
-import SvelButton from '@svelement-ui/button/src/svel-button.svelte';
-export { default as SvelCol } from '@svelement-ui/col';
-export { default as SvelContainer } from '@svelement-ui/container';
-export { default as SvelFooter } from '@svelement-ui/footer';
-export { default as SvelHeader } from '@svelement-ui/header';
-export { default as SvelMain } from '@svelement-ui/main';
-export { default as SvelRow } from '@svelement-ui/row';
+import SvelAside from '@svelement-ui/aside';
+import SvelButton from '@svelement-ui/button';
+import SvelCol from '@svelement-ui/col';
+import SvelContainer from '@svelement-ui/container';
+import SvelFooter from '@svelement-ui/footer';
+import SvelHeader from '@svelement-ui/header';
+import SvelMain from '@svelement-ui/main';
+import SvelRow from '@svelement-ui/row';
 
-export { SvelAside, SvelButton };
+export { SvelAside, SvelButton, SvelCol, SvelContainer, SvelFooter, SvelHeader, SvelMain, SvelRow };