+++ packages/header/src/lib/index.ts
@@ -1,2 +1,0 @@
-import SvelHeader from './svel-header.svelte';
-export default SvelHeader;