+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.5](https://github.com/koory1st/svelement-ui/compare/v0.8.4...v0.8.5) (2023-08-20)
+
+
+### Bug Fixes
+
+* **button:** change package ([1e9ad3e](https://github.com/koory1st/svelement-ui/commit/1e9ad3e46f6dfe5f5634da99573e6c429da50f8c))
+* **col:** simplify package ([41479ac](https://github.com/koory1st/svelement-ui/commit/41479ac6167a9bd785c511861d089941f22742b9))
+* **container:** simplify package ([d35abaf](https://github.com/koory1st/svelement-ui/commit/d35abafb2b745490c6107cb9de7e6c6dc43b99c8))
+* **footer:** simplify package ([55e273b](https://github.com/koory1st/svelement-ui/commit/55e273b85a78296b76036e66685cc60cf60b32f2))
+* **header:** simplify package ([91eef62](https://github.com/koory1st/svelement-ui/commit/91eef620c713c430c44f1d23f94421aa913ebc25))
+* **main:** simplify package ([da30a9d](https://github.com/koory1st/svelement-ui/commit/da30a9d55aec92533633dea36ad2e8e7d09f4bc8))
+* **row:** simplify package ([371f2ba](https://github.com/koory1st/svelement-ui/commit/371f2ba1c4e306af77d9aad8fcfe88a4d2340764))
+
+
+
+
+
 ## [0.8.4](https://github.com/koory1st/svelement-ui/compare/v0.8.3...v0.8.4) (2023-08-18)